requirements.txt
Django>=2.2.17,<3.0.0
djangorestframework>=3.12.2,<3.13.0
psycopg2-binary>=2.8.6,<2.9.0
djangorestframework_simplejwt>=4.6.0,<4.7.0

flake8>=3.8.4,<3.9.0